Renamed modules to sampling.py and __main__.py
