+++ pdmm/sampling.py
